Indexing almost done

+++ src/com/company/Posting.java
@@ -5,11 +5,11 @@ import java.util.ArrayList;
 /* Document Posting Class */
 public class Posting {
 
-    Long DocId;    // Document Id
-    Long termFreq;  // Term Frequency
-    ArrayList<Long> pos;  // List of Positions in the Document
+    Integer DocId;    // Document Id
+    Integer termFreq;  // Term Frequency
+    ArrayList<Integer> pos;  // List of Positions in the Document
 
-    public Posting(Long docId, Long termFreq) {
+    public Posting(Integer docId, Integer termFreq) {
         DocId = docId;
         this.termFreq = termFreq;
     }
@@ -18,28 +18,28 @@ public class Posting {
 
     }
 
-    public Long getDocId() {
+    public Integer getDocId() {
 
         return DocId;
     }
 
-    public void setDocId(Long docId) {
+    public void setDocId(Integer docId) {
         DocId = docId;
     }
 
-    public Long getTermFreq() {
+    public Integer getTermFreq() {
         return termFreq;
     }
 
-    public void setTermFreq(Long termFreq) {
+    public void setTermFreq(int termFreq) {
         this.termFreq = termFreq;
     }
 
-    public ArrayList<Long> getPos() {
+    public ArrayList<Integer> getPos() {
         return pos;
     }
 
-    public void setPos(ArrayList<Long> pos) {
+    public void setPos(ArrayList<Integer> pos) {
         this.pos = pos;
     }
 }
